# docs/openapi-novel-system.yml
openapi: 3.0.0
info:
  title: 超长篇小说AI创作系统 API
  description: 超长篇小说AI创作系统 (v16.0) 的API接口文档
  version: 16.0.0
  contact:
    name: Novel AI System Support
    email: support@novel-ai-system.com

servers:
  - url: http://localhost:3000
    description: 本地开发服务器
  - url: https://api.novel-ai-system.com
    description: 生产服务器

paths:
  /api/projects:
    post:
      summary: 创建新项目
      description: 初始化一个新的小说项目，创建目录结构和基础配置文件
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - chapterCount
              properties:
                name:
                  type: string
                  description: 项目名称
                  example: "我的玄幻小说"
                chapterCount:
                  type: integer
                  description: 章节数量
                  example: 100
                genre:
                  type: string
                  description: 小说类型
                  example: "玄幻"
      responses:
        '201':
          description: 项目创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    get:
      summary: 获取项目列表
      description: 获取所有已创建的项目
      responses:
        '200':
          description: 项目列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/projects/{projectId}:
    get:
      summary: 获取项目详情
      description: 获取指定项目的信息
      parameters:
        - name: projectId
          in: path
          required: true
          description: 项目ID
          schema:
            type: string
      responses:
        '200':
          description: 项目详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          description: 项目不存在
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      summary: 更新项目配置
      description: 更新项目的基本配置信息
      parameters:
        - name: projectId
          in: path
          required: true
          description: 项目ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdate'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          description: 项目不存在
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/projects/{projectId}/outline:
    post:
      summary: 生成小说大纲
      description: 基于项目设定生成小说大纲
      parameters:
        - name: projectId
          in: path
          required: true
          description: 项目ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                genre:
                  type: string
                  description: 小说类型
                  example: "玄幻"
                protagonist:
                  type: string
                  description: 主角设定
                  example: "废柴少年"
                mainConflict:
                  type: string
                  description: 主要冲突
                  example: "宗门试炼中的生存挑战"
      responses:
        '200':
          description: 大纲生成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Outline'
        '400':
          description: 请求参数错误
        '404':
          description: 项目不存在
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/projects/{projectId}/chapters:
    post:
      summary: 批量创建章节
      description: 根据大纲和设定批量创建多个章节
      parameters:
        - name: projectId
          in: path
          required: true
          description: 项目ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - startChapter
                - endChapter
              properties:
                startChapter:
                  type: integer
                  description: 起始章节号
                  example: 1
                endChapter:
                  type: integer
                  description: 结束章节号
                  example: 10
      responses:
        '200':
          description: 章节创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  createdChapters:
                    type: array
                    items:
                      $ref: '#/components/schemas/Chapter'
                  summary:
                    type: string
                    description: 创建摘要
        '400':
          description: 请求参数错误
        '404':
          description: 项目不存在
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/projects/{projectId}/chapters/{chapterNumber}:
    get:
      summary: 获取章节内容
      description: 获取指定章节的内容
      parameters:
        - name: projectId
          in: path
          required: true
          description: 项目ID
          schema:
            type: string
        - name: chapterNumber
          in: path
          required: true
          description: 章节号
          schema:
            type: integer
      responses:
        '200':
          description: 章节内容
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chapter'
        '404':
          description: 章节或项目不存在
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      summary: 更新章节内容
      description: 更新指定章节的内容
      parameters:
        - name: projectId
          in: path
          required: true
          description: 项目ID
          schema:
            type: string
        - name: chapterNumber
          in: path
          required: true
          description: 章节号
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: 章节标题
                content:
                  type: string
                  description: 章节内容
                  format: markdown
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chapter'
        '404':
          description: 章节或项目不存在
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/projects/{projectId}/settings:
    get:
      summary: 获取项目设定
      description: 获取项目的所有设定信息（角色、世界观、力量体系等）
      parameters:
        - name: projectId
          in: path
          required: true
          description: 项目ID
          schema:
            type: string
      responses:
        '200':
          description: 项目设定
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectSettings'
        '404':
          description: 项目不存在
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      summary: 更新项目设定
      description: 更新项目的所有设定信息
      parameters:
        - name: projectId
          in: path
          required: true
          description: 项目ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectSettings'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectSettings'
        '404':
          description: 项目不存在
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/token-usage:
    get:
      summary: 获取令牌使用情况
      description: 获取当前会话的令牌使用情况
      responses:
        '200':
          description: 令牌使用情况
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenUsage'
        '500':
          description: 服务器错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Project:
      type: object
      properties:
        id:
          type: string
          description: 项目唯一标识符
        name:
          type: string
          description: 项目名称
        chapterCount:
          type: integer
          description: 章节数量
        genre:
          type: string
          description: 小说类型
        status:
          type: string
          description: 项目状态
          enum: [initialized, outline_created, chapters_writing, completed]
        createdAt:
          type: string
          format: date-time
          description: 创建时间
        currentChapter:
          type: integer
          description: 当前创作章节号
        totalWords:
          type: integer
          description: 总字数
      required:
        - id
        - name
        - chapterCount
        - genre
        - status
        - createdAt

    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
          description: 项目名称
        genre:
          type: string
          description: 小说类型
        chapterCount:
          type: integer
          description: 章节数量
        status:
          type: string
          description: 项目状态

    Outline:
      type: object
      properties:
        chapters:
          type: array
          items:
            type: object
            properties:
              number:
                type: integer
                description: 章节号
              title:
                type: string
                description: 章节标题
              summary:
                type: string
                description: 章节概要
              keyEvents:
                type: array
                items:
                  type: string
                description: 关键事件列表

    Chapter:
      type: object
      properties:
        number:
          type: integer
          description: 章节号
        title:
          type: string
          description: 章节标题
        content:
          type: string
          description: 章节内容
          format: markdown
        wordCount:
          type: integer
          description: 字数
        createdAt:
          type: string
          format: date-time
          description: 创建时间
        updatedAt:
          type: string
          format: date-time
          description: 更新时间

    ProjectSettings:
      type: object
      properties:
        characters:
          type: object
          description: 角色设定
          properties:
            protagonist:
              $ref: '#/components/schemas/Character'
            supporting:
              type: array
              items:
                $ref: '#/components/schemas/Character'
            antagonists:
              type: array
              items:
                $ref: '#/components/schemas/Character'
        worldview:
          $ref: '#/components/schemas/Worldview'
        powerSystem:
          $ref: '#/components/schemas/PowerSystem'
        foreshadows:
          type: array
          items:
            $ref: '#/components/schemas/Foreshadow'

    Character:
      type: object
      properties:
        name:
          type: string
          description: 角色名称
        description:
          type: string
          description: 角色描述
        personality:
          type: string
          description: 性格特点
        abilities:
          type: array
          items:
            type: string
          description: 能力列表
        development:
          type: array
          items:
            type: string
          description: 角色发展轨迹

    Worldview:
      type: object
      properties:
        setting:
          type: string
          description: 世界设定
        rules:
          type: object
          description: 世界规则
        cultures:
          type: array
          items:
            type: string
          description: 文化列表
        geography:
          type: string
          description: 地理环境
        history:
          type: string
          description: 历史背景
        magicSystem:
          type: object
          description: 魔法体系
        technologyLevel:
          type: string
          description: 科技水平
        socialStructure:
          type: string
          description: 社会结构

    PowerSystem:
      type: object
      properties:
        name:
          type: string
          description: 力量体系名称
        levels:
          type: array
          items:
            type: string
          description: 等级列表
        requirements:
          type: string
          description: 晋升要求
        limitations:
          type: string
          description: 限制条件
        acquisitionMethod:
          type: string
          description: 获得方式
        trainingProcess:
          type: string
          description: 修炼过程
        powerSources:
          type: array
          items:
            type: string
          description: 力量来源

    Foreshadow:
      type: object
      properties:
        description:
          type: string
          description: 伏笔描述
        chapter:
          type: integer
          description: 所在章节
        resolved:
          type: boolean
          description: 是否已回收

    TokenUsage:
      type: object
      properties:
        used:
          type: integer
          description: 已使用令牌数
        available:
          type: integer
          description: 可用令牌数
        limit:
          type: integer
          description: 令牌总数限制
        safetyMargin:
          type: integer
          description: 安全余量

    Error:
      type: object
      properties:
        code:
          type: integer
          description: 错误代码
        message:
          type: string
          description: 错误消息
        details:
          type: object
          description: 错误详情
      required:
        - code
        - message